requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.16.0
channels==3.0.4
charset-normalizer==3.3.0
cryptography==41.0.4
defusedxml==0.8.0rc2
Django==4.2.5
django-allauth==0.57.0
django-jazzmin==2.6.0
django-leaflet==0.29.0
idna==3.4
oauthlib==3.2.2
Pillow==10.1.0
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
razorpay==1.4.1
reportlab==4.0.7
requests==2.31.0
requests-oauthlib==1.3.1
social-auth-app-django==5.3.0
social-auth-core==4.4.2
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.6
